+++ test/writing_outputs/test_zone_no_resources.jl
@@ -19,16 +19,18 @@ end
 
 function prepare_costs_true()
     df = DataFrame(
-        ["cTotal" 5.177363815260002e12 4.027191550200002e12 1.1501722650599993e12;
-         "cFix" 0.0 0.0 0.0;
-         "cVar" 5.849292224195126e-8 0.0 5.849292224195126e-8;
-         "cFuel" 0.0 0.0 0.0;
-         "cNSE" 5.177363815260002e12 4.027191550200002e12 1.1501722650599993e12;
-         "cStart" 0.0 0.0 0.0;
-         "cUnmetRsv" 0.0 0.0 0.0;
-         "cNetworkExp" 0.0 0.0 0.0;
-         "cUnmetPolicyPenalty" 0.0 0.0 0.0;
-         "cCO2" 0.0 0.0 0.0],
+        ["cTotal" 1.463768296693398e12 1.4588919221877012e12 4.840797505702658e9;
+        "cFix" 4.727103377348504e9 0.0 4.727103377348504e9;
+        "cVar" 2.3347293240589075e7 0.0 2.3347293240589064e7;
+        "cFuel" 4.4718223470249504e7 0.0 4.4718223470249504e7;
+        "cNSE" 1.4588919221877012e12 1.4588919221877012e12 0.0;
+        "cStart" 4.5628611643315025e7 0.0 4.5628611643315025e7;
+        "cUnmetRsv" 0.0 0.0 0.0;
+        "cNetworkExp" 3.5577e7 0.0 0.0;
+        "cUnmetPolicyPenalty" 0.0 0.0 0.0;
+        "cCO2" 0.0 0.0 0.0;
+        "cInv" 3.4603915632743273e9 0.0 3.4603915632743273e9;
+        "cFom" 1.266711814074177e9 0.0 1.266711814074177e9],
         [:Costs, :Total, :Zone1, :Zone2])
 
     df[!, :Costs] = convert(Vector{String}, df[!, :Costs])
@@ -40,7 +42,7 @@ end
 
 function test_case()
     test_path = joinpath(@__DIR__, "zone_no_resources")
-    obj_true = 5.1773638153e12
+    obj_true = 1.463768296693398e12
     costs_true = prepare_costs_true()
 
     # Define test setup
@@ -60,11 +62,12 @@ function test_case()
     optimal_tol = optimal_tol_rel * obj_test  # Convert to absolute tolerance
 
     # Test the objective value
-    test_result = @test obj_test≈obj_true atol=optimal_tol
+    @test obj_test≈obj_true atol=optimal_tol
 
     # Test the costs
     costs_test = prepare_costs_test(test_path, inputs, genx_setup, EP)
-    test_result = @test costs_test[!, Not(:Costs)] ≈ costs_true[!, Not(:Costs)]
+    @test costs_test[!, Not(:Costs)] ≈ costs_true[!, Not(:Costs)]
+    @test Vector(costs_test[2, 2:end]) ≈ (Vector(costs_true[end-1, 2:end]) + Vector(costs_true[end, 2:end]))
 
     # Remove the costs file
     rm(joinpath(test_path, "costs.csv"))